AK: StringView uses WeakPtr instead of raw pointer to StringImpl
This allows more safety in terms of accessing string view characters
when the memory already freed. When using a WeakPtr to the StringImpl,
we can check if the impl already has gone and assure freed memory is
not being accessed anymore.
New constructors for the StringView allow to hand over the impl, and
Strings substring_view method and StringViews own substring_view*
methods hand over the impl to the StringView constructors.
StringViews created from plain C character are not affected.

Needed some const_casts to allow the creation of the WeakPtr and
we don't wont to loose the constness of the String or FlyString
handed over in the constructor.